requirements.txt
mlflow==2.9.0 
langchain==0.0.344 
databricks-vectorsearch==0.22 
cloudpickle==2.2.1 
databricks-sdk==0.12.0 
cloudpickle==2.2.1 
pydantic==2.5.2 
transformers==4.34.0 
databricks-genai-inference==0.1.1
textstat==0.7.3 
tiktoken==0.5.1 
evaluate==0.4.1
torch==2.0.1
pandas==1.5.3
numpy==1.23.5
databricks-connect==14.2.1
pyyaml